discord-ai/requirements.txt
discord
python-dotenv
transformers
tensorflow
torch
langchain
openai
faiss-cpu
requests
tiktoken